combine the xds-translator and xds-server runner into one (#6586)
* combine the xds-translator and xds-server into one xds runner

* primarily to reduce memory and convergence time
